requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-crontab==0.7.1
djangorestframework==3.7.7
future==0.16.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
py4j==0.10.6
pyspark==2.3.0
python-twitter==3.4.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
